+++ Sources/xcresultparser/CoberturaCoverageConverter.swift
@@ -35,53 +35,6 @@ enum JSONParseError: Error {
     case convertError(code: Int, message: String)
 }
 
-// Subrange information struct
-struct Subrange: Decodable {
-    let column: Int
-    let executionCount: Int
-    let length: Int
-}
-
-// LineDetail information struct
-struct LineDetail: Decodable {
-    let isExecutable: Bool
-    let line: Int
-    let executionCount: Int?
-    let subranges: [Subrange]?
-}
-
-// FileCoverage information struct
-struct FileCoverage: Decodable {
-    let files: [String: [LineDetail]]
-
-    // Custom initializer to handle the top-level dictionary
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        var filesDict = [String: [LineDetail]]()
-
-        for key in container.allKeys {
-            let keyString = key.stringValue
-            let lineDetails = try container.decode([LineDetail].self, forKey: key)
-            filesDict[keyString] = lineDetails
-        }
-        self.files = filesDict
-    }
-
-    private struct CodingKeys: CodingKey {
-        var stringValue: String
-        var intValue: Int?
-
-        init?(stringValue: String) {
-            self.stringValue = stringValue
-            self.intValue = nil
-        }
-
-        init?(intValue: Int) {
-            return nil
-        }
-    }
-}
-
 public class CoberturaCoverageConverter: CoverageConverter, XmlSerializable {
 
     public var xmlString: String {
@@ -112,54 +65,28 @@ public class CoberturaCoverageConverter: CoverageConverter, XmlSerializable {
         rootElement.addChild(packagesElement)
 
         // Get the xccov results as a JSON.
-        var arguments = ["xccov", "view"]
-        if resultFile.url.pathExtension == "xcresult" {
-            arguments.append("--archive")
-        }
-        arguments.append("--json")
-        arguments.append(resultFile.url.path)
-        let coverageData = try Shell.execute(program: "/usr/bin/xcrun", with: arguments)
-        let resultsString = String(decoding: coverageData, as: UTF8.self)
-        guard let jsonData = resultsString.data(using: String.Encoding.utf8) else {
-            writeToStdError("Failed to convert to Data")
-            throw JSONParseError.convertError(code: 0, message: "Failed to convert to Data object")
-        }
-        guard let coverageJson = try? JSONDecoder().decode(FileCoverage.self, from: jsonData) else {
-            writeToStdError("Failed to convert to JSON")
-            throw JSONParseError.convertError(code: 0, message: "Failed to convert to JSON Object")
-        }
+        let coverageJson = try getCoverageDataAsJSON()
 
-        let fileInfoSemaphore = DispatchSemaphore(value: 1)
         var fileInfo: [FileInfo] = []
-        let queue = OperationQueue()
-        queue.maxConcurrentOperationCount = ProcessInfo.processInfo.processorCount //Deadlock if this is = 1
-        queue.qualityOfService = .userInitiated
         for (fileName, value) in coverageJson.files {
-            let op = BlockOperation {
-                var fileLines = [LineInfo]() // This will store information about each line.
-
-                for lineData in value {
-                    let lineNum = lineData.line
-                    guard var covered = lineData.executionCount else {
-                        continue
-                    }
-                    // If the line coverage count is a MAX_INT, just set it to 1
-                    if covered == Int.max {
-                        covered = 1
-                    }
-                    let line = LineInfo(lineNumber: String(lineNum), coverage: covered)
-                    fileLines.append(line)
+            var fileLines = [LineInfo]() // This will store information about each line.
+            
+            for lineData in value {
+                let lineNum = lineData.line
+                guard var covered = lineData.executionCount else {
+                    continue
                 }
-
-                let fileInfoInst = FileInfo(path: self.relativePath(for: fileName, relativeTo: self.projectRoot), lines: fileLines)
-                fileInfoSemaphore.wait()
-                fileInfo.append(fileInfoInst)
-                fileInfoSemaphore.signal()
+                // If the line coverage count is a MAX_INT, just set it to 1
+                if covered == Int.max {
+                    covered = 1
+                }
+                let line = LineInfo(lineNumber: String(lineNum), coverage: covered)
+                fileLines.append(line)
             }
-            queue.addOperation(op)
+            
+            let fileInfoInst = FileInfo(path: self.relativePath(for: fileName, relativeTo: self.projectRoot), lines: fileLines)
+            fileInfo.append(fileInfoInst)
         }
-        // This will block until all our operation have compleated (or been canceled)
-        queue.waitUntilAllOperationsAreFinished()
         // Sort files to avoid duplicated packages
         fileInfo.sort { $0.path > $1.path }
         
@@ -231,24 +158,7 @@ public class CoberturaCoverageConverter: CoverageConverter, XmlSerializable {
         fatalError("DTD could not be constructed")
     }
     
-    private func fileCoverage(for file: String, relativeTo projectRoot: String) throws -> FileInfo {
-        let coverageData = try coverageForFile(path: file)
-        var file = FileInfo(path: relativePath(for: file, relativeTo: projectRoot), lines: [])
-        let nsrange = NSRange(coverageData.startIndex..<coverageData.endIndex,
-                              in: coverageData)
-        coverageRegexp!.enumerateMatches(in: coverageData, options: [], range: nsrange) { match, flags, stop in
-            guard let match = match else { return }
-            
-            let lineNumber = coverageData.text(in: match.range(at: 1))
-            let coverage = coverageData.text(in: match.range(at: 2))
-            
-            let line = LineInfo(lineNumber: lineNumber, coverage: Int(coverage) ?? 0)
-            file.lines.append(line)
-        }
-        return file
-    }
-    
-    func makeRootElement() -> XMLElement {
+    private func makeRootElement() -> XMLElement {
         // TODO some of these values are B.S. - figure out how to calculate, or better to omit if we don't know?
         let testAction = invocationRecord.actions.first { $0.schemeCommandName == "Test" }
         let timeStamp = (testAction?.startedTime.timeIntervalSince1970) ?? Date().timeIntervalSince1970
@@ -267,17 +177,17 @@ public class CoberturaCoverageConverter: CoverageConverter, XmlSerializable {
     }
 }
 
-struct LineInfo {
+private struct LineInfo {
     let lineNumber: String
     let coverage: Int
 }
 
-struct FileInfo {
+private struct FileInfo {
     let path: String
     var lines: [LineInfo]
 }
 
-extension XMLNode {
+private extension XMLNode {
     static func nodeAttribute(withName name: String, stringValue value: String) -> XMLNode {
         guard let attribute = XMLNode.attribute(withName: name, stringValue: value) as? XMLNode else {
             return XMLNode()