+++ Sources/xcresultparser/Models/Coverage/Subrange.swift
@@ -1,0 +1,14 @@
+//
+//  Subrange.swift
+//
+//  Created by maxwell-legrand on 15.05.24.
+//
+
+import Foundation
+
+// Subrange information struct
+struct Subrange: Decodable {
+    let column: Int
+    let executionCount: Int
+    let length: Int
+}